feat!: rename, reorganize and improve key_reader
Split `key_reader.ts` to multiple modules inside `input_reader/``.
`input_reader/decoders` contains logic for decoding given buffer/character code to events.

Replace `readKeypresses()`` with `emitInputEvents()` which directly emits input from given stdin with parsed events.

`key_reader.ts` related types have been moved from `src/types.ts` to `src/input_reader/types.ts`.

`"mousePress"` event has been split to three different events: "mouseEvent", "mousePress", "mouseScroll".
During testing I was checking on handling "mouseMove" but it has visible cpu-usage while just moving mouse so I've abandoned it, I might introduce it behind opt-in option.

Some general optimizations that have been done include:
 - instead of splitting and then mapping codes they're now parsed via string slicing
 - instead of using `TextDecoder.decode` `decodeBuffer()` uses `fastSmallDecode()` function which performs up to 70% better on small strings.
 - mouse decoder now re-uses two objects instead of creating new ones constantly (should be free of GC) (this should be done with keypress too!)
