+++ mod.ts
@@ -3,9 +3,12 @@ export * from "./src/component.ts";
 export * from "./src/controls.ts";
 export * from "./src/event_emitter.ts";
 export * from "./src/input.ts";
-export * from "./src/key_reader.ts";
 export * from "./src/theme.ts";
 export * from "./src/tui.ts";
 export * from "./src/types.ts";
+
 export * from "./src/canvas/mod.ts";
+
 export * from "./src/utils/mod.ts";
+
+export * from "./src/input_reader/mod.ts";
